POS: Finished fetching and population of customer information, CRUD for shoping list and population of product items on the dropdown. Discount and fees sync on ordersummary section. dynamic Total Amount on the order summary.

+++ ABC.Shared/Models/CustomerBasicInfo.cs
@@ -5,6 +5,6 @@ namespace ABC.Shared.Models
         public string Id { get; set; } = String.Empty;
         public string FirstName { get; set; } = String.Empty;
         public string LastName { get; set; }= String.Empty;
-        public int ContactNumber { get; set; }= 0;
+        public long ContactNumber { get; set; }= 0;
     }
 }
